+++ cmd/devdash/devdash.go
@@ -17,7 +17,6 @@ import (
 var debug *bool
 
 func main() {
-	// TODO add mor commands + better management of them (cobra?)
 	file := flag.String("config", ".devdash.yml", "The config file")
 	debug = flag.Bool("debug", false, "Debug mode")
 	term := flag.Bool("term", false, "Display terminal dimensions")
@@ -41,6 +40,7 @@ func main() {
 	run(*file, tui)()
 
 	if _, err := os.Stat(*file); os.IsNotExist(err) {
+		tui.AddKQuit("C-c")
 		internal.DisplayNoFile(tui)
 		err := tui.AddCol("5")
 		if err != nil {
@@ -132,6 +132,22 @@ func run(file string, tui *internal.Tui) func() {
 					travisService.Token,
 				)
 				project.WithTravisCI(travisWidget)
+			}
+
+			feedlyService := p.Services.Feedly
+			if !feedlyService.empty() {
+				feedlyService := internal.NewFeedlyWidget(
+					feedlyService.Address,
+				)
+				project.WithFeedly(feedlyService)
+			}
+
+			gitService := p.Services.Git
+			if !gitService.empty() {
+				gitService := internal.NewGitWidget(
+					gitService.Path,
+				)
+				project.WithGit(gitService)
 			}
 
 			project.Render(*debug)